% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_n}
\alias{add_n}
\title{Add \code{n} to a Number}
\usage{
add_n(x, n)
}
\arguments{
\item{x}{number to which \code{n} should be added}

\item{n}{number to add to \code{x}}
}
\value{
\code{x+n}
}
\description{
Add \code{n} to a Number
}
\seealso{
Other helpers: 
\code{\link{ret_err}()},
\code{\link{ret_null}()},
\code{\link{sleep_for}()},
\code{\link{square}()},
\code{\link{timestwo}()}
}
\concept{helpers}
