% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{Err}
\alias{Err}
\title{Create an \code{Err} instance}
\usage{
Err(...)
}
\arguments{
\item{...}{error messages to be contained}
}
\value{
an \code{Err} instance
}
\description{
Create an \code{Err} instance
}
\seealso{
Other result: 
\code{\link{Ok}()},
\code{\link{is_err}()},
\code{\link{is_ok}()}
}
\concept{result}
