% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap.R
\name{fmap}
\alias{fmap}
\title{fmap: unwrap, apply function, and rewrap.}
\usage{
fmap(.m, .f, ...)
}
\arguments{
\item{.m}{The monad}

\item{.f}{A function, formula, or atomic vector.

  If a \strong{function}, it is used as is.

  If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a
  function with two arguments, \code{.x} or \code{.} and \code{.y}. This
  allows you to create very compact anonymous functions with up to
  two inputs.

  If \strong{character} or \strong{integer vector}, e.g. \code{"y"}, it
  is converted to an extractor function, \code{function(x) x[["y"]]}. To
  index deeply into a nested list, use multiple values; \code{c("x", "y")}
  is equivalent to \code{z[["x"]][["y"]]}. You can also set \code{.null}
  to set a default to use instead of \code{NULL} for absent components.}

\item{...}{Additonal arguments passed on to \code{.f}.}
}
\description{
Any object with a \code{fmap} method is called a \code{functor}.
}
\examples{
# Functions are a functor
add1 <- function(x) x + 1
add2 <- function(x) x + 2

add3 <- add1 \%>>\% add2
add3(10)
}

