% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{\%>>=\%}
\alias{\%>>=\%}
\title{flatMap}
\usage{
lhs \%>>=\% rhs
}
\arguments{
\item{lhs}{left side of infix}

\item{rhs}{right side of infix}
}
\value{
\code{rhs(lhs)} accounting in the (monad) context of \code{lhs}
}
\description{
Applies the \code{rhs} as an argument to \code{lhs} in the context of the \code{lhs} monad.
}
\details{
Essentially calls \code{lhs$bind(rhs)} which for most cases, will return \code{rhs(lhs)}
with whatever context the monad provides.
}
