% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%>>\%}
\alias{\%>>\%}
\title{Infix fmap: allows you to apply an arbitrary function to a wrapped object.}
\usage{
lhs \%>>\% rhs
}
\arguments{
\item{lhs}{A monad}

\item{rhs}{Call to invoke.}
}
\description{
This works like magrittr's \code{\%>\%}, but uses fmap to unwrap, apply,
and then rewrap the object.
}
\examples{
if (require("shiny")) {
options(shiny.suppressMissingContextError = TRUE)
r <- reactive(iris)
r2 <- r \%>>\% head(5)
r2()
}
}

