% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{is_err}
\alias{is_err}
\title{Test for \code{Err}}
\usage{
is_err(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical; whether \code{x} is an \code{Err} instance
}
\description{
Test for \code{Err}
}
\seealso{
Other result: 
\code{\link{Err}()},
\code{\link{Ok}()},
\code{\link{is_ok}()}
}
\concept{result}
