% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{is_just}
\alias{is_just}
\title{Test for \code{Just}}
\usage{
is_just(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical; whether \code{x} is a \code{Just} instance
}
\description{
Test for \code{Just}
}
\seealso{
Other maybe: 
\code{\link{Just}()},
\code{\link{Nothing}()},
\code{\link{is_nothing}()}
}
\concept{maybe}
