% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{is_nothing}
\alias{is_nothing}
\title{Test for \code{Nothing}}
\usage{
is_nothing(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical; whether \code{x} is a \code{Nothing} instance
}
\description{
Test for \code{Nothing}
}
\seealso{
Other maybe: 
\code{\link{Just}()},
\code{\link{Nothing}()},
\code{\link{is_just}()}
}
\concept{maybe}
