% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{is_ok}
\alias{is_ok}
\title{Test for \code{Ok}}
\usage{
is_ok(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical; whether \code{x} is an \code{Ok} instance
}
\description{
Test for \code{Ok}
}
\seealso{
Other result: 
\code{\link{Err}()},
\code{\link{Ok}()},
\code{\link{is_err}()}
}
\concept{result}
