% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{Just}
\alias{Just}
\title{Create a \code{Just} instance}
\usage{
Just(x)
}
\arguments{
\item{x}{value to be contained}
}
\value{
a \code{Just} instance
}
\description{
Create a \code{Just} instance
}
\seealso{
Other maybe: 
\code{\link{Nothing}()},
\code{\link{is_just}()},
\code{\link{is_nothing}()}
}
\concept{maybe}
