% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{List}
\alias{List}
\title{List Monad}
\description{
Creates a \code{List} monad instance which can be used with the flatMap
operator \verb{\%>>=\%}. See the \verb{Monads in R} vignette for more details.

This container, when flatMapped, unpacks the contained values, concatenates
them, passes this to the next expression, then rewraps the result in a new
\code{List}.
}
\seealso{
\code{\link[=times]{times()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{value contained in the instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ListMonad-new}{\code{List$new()}}
\item \href{#method-ListMonad-bind}{\code{List$bind()}}
\item \href{#method-ListMonad-clone}{\code{List$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMonad-new"></a>}}
\if{latex}{\out{\hypertarget{method-ListMonad-new}{}}}
\subsection{Method \code{new()}}{
create a new \code{List} instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{List$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{values to be contained}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{List} instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMonad-bind"></a>}}
\if{latex}{\out{\hypertarget{method-ListMonad-bind}{}}}
\subsection{Method \code{bind()}}{
bind when using flatMap (\verb{\%>>=\%})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{List$bind(f, expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{function to apply}

\item{\code{expr}}{expression of the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMonad-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ListMonad-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{List$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
