% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{listM}
\alias{listM}
\title{Create a new \code{List} instance}
\usage{
listM(...)
}
\arguments{
\item{...}{values to be contained}
}
\value{
a \code{List} monad instance
}
\description{
Create a new \code{List} instance
}
