% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOf.R
\name{listOf}
\alias{listOf}
\title{A list where all elements are the same class.}
\usage{
listOf(...)
}
\arguments{
\item{...}{Elements to combine into a list. All elements must be the
same type.}
}
\description{
A list where all elements are the same class.
}
\examples{
x <- listOf(1L, 2L, 4:10)
double <- function(x) x * 2
x \%>>\% double()

x \%>>\% function(x) list(x, x)
x \%>+\% function(x) list(x, x)
}

