% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Logger}
\alias{Logger}
\title{Logger Monad}
\description{
Creates a \code{Logger} monad instance which can be used with the flatMap
operator \verb{\%>>=\%}. See the \verb{Monads in R} vignette for more details.

This container, when flatMapped, records the expressions used at each step,
and these can be recovered from the object, potentially for re-evaluation.
}
\seealso{
\code{\link[=rerun]{rerun()}}, \code{\link[=logger_log]{logger_log()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{value contained in the instance}

\item{\code{log}}{log of the evaluated expressions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoggerMonad-new}{\code{Logger$new()}}
\item \href{#method-LoggerMonad-bind}{\code{Logger$bind()}}
\item \href{#method-LoggerMonad-clone}{\code{Logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggerMonad-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoggerMonad-new}{}}}
\subsection{Method \code{new()}}{
create a new \code{Logger} instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value to be contained}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{Logger} instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggerMonad-bind"></a>}}
\if{latex}{\out{\hypertarget{method-LoggerMonad-bind}{}}}
\subsection{Method \code{bind()}}{
bind when using flatMap (\verb{\%>>=\%})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$bind(f, expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{function to apply}

\item{\code{expr}}{expression of the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggerMonad-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoggerMonad-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
