% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_log}
\alias{logger_log}
\title{Extact the log}
\usage{
logger_log(x, as_text = FALSE)
}
\arguments{
\item{x}{a \code{Logger} monad instance}

\item{as_text}{(logical) should the log be returned as text?}
}
\value{
if \code{as_text}, a string, otherwise the log is printed to console
}
\description{
Extact the log
}
\seealso{
Other logger: 
\code{\link{rerun}()}
}
\concept{logger}
