% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{loggerM}
\alias{loggerM}
\title{Create a new \code{Logger} instance}
\usage{
loggerM(value)
}
\arguments{
\item{value}{value to be contained}
}
\value{
a \code{Logger} monad instance
}
\description{
Create a new \code{Logger} instance
}
