% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{Maybe}
\alias{Maybe}
\title{Maybe Monad}
\description{
Creates a \code{Maybe} monad instance which can be used with the flatMap
operator \verb{\%>>=\%}. See the \verb{Monads in R} vignette for more details.

This container, when flatMapped, always returns a value, either of type
\code{Nothing} (if an error was raised or the result was \code{NULL}), or \code{Just}.
}
\seealso{
\code{\link[=Nothing]{Nothing()}}, \code{\link[=is_nothing]{is_nothing()}}, \code{\link[=Just]{Just()}}, \code{\link[=is_just]{is_just()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{value contained in the instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MaybeMonad-new}{\code{Maybe$new()}}
\item \href{#method-MaybeMonad-bind}{\code{Maybe$bind()}}
\item \href{#method-MaybeMonad-clone}{\code{Maybe$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaybeMonad-new"></a>}}
\if{latex}{\out{\hypertarget{method-MaybeMonad-new}{}}}
\subsection{Method \code{new()}}{
create a new \code{Maybe} instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$new(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value to be contained}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{Maybe} instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaybeMonad-bind"></a>}}
\if{latex}{\out{\hypertarget{method-MaybeMonad-bind}{}}}
\subsection{Method \code{bind()}}{
bind when using flatMap (\verb{\%>>=\%})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$bind(f, expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{function to apply}

\item{\code{expr}}{expression of the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaybeMonad-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MaybeMonad-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Maybe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
