% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybeM}
\alias{maybeM}
\title{Create a new \code{Maybe} instance}
\usage{
maybeM(value)
}
\arguments{
\item{value}{value to be contained}
}
\value{
a \code{Maybe} monad instance
}
\description{
Create a new \code{Maybe} instance
}
