% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{Nothing}
\alias{Nothing}
\title{Create an \code{Nothing} instance}
\usage{
Nothing()
}
\value{
a \code{Nothing} instance
}
\description{
Create an \code{Nothing} instance
}
\seealso{
Other maybe: 
\code{\link{Just}()},
\code{\link{is_just}()},
\code{\link{is_nothing}()}
}
\concept{maybe}
