% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{Ok}
\alias{Ok}
\title{Create an \code{Ok} instance}
\usage{
Ok(x)
}
\arguments{
\item{x}{value to be contained}
}
\value{
an \code{Ok} instance
}
\description{
Create an \code{Ok} instance
}
\seealso{
Other result: 
\code{\link{Err}()},
\code{\link{is_err}()},
\code{\link{is_ok}()}
}
\concept{result}
