% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rerun}
\alias{rerun}
\title{(Re)-evaluate the logged commands as a regular pipeline}
\usage{
rerun(x)
}
\arguments{
\item{x}{a \code{Logger} monad instance}
}
\value{
the evaluated output of the logged pipeline
}
\description{
(Re)-evaluate the logged commands as a regular pipeline
}
\seealso{
Other logger: 
\code{\link{logger_log}()}
}
\concept{logger}
