% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{resultM}
\alias{resultM}
\title{Create a new \code{Result} instance}
\usage{
resultM(value)
}
\arguments{
\item{value}{value to be contained}
}
\value{
a \code{Result} monad instance
}
\description{
Create a new \code{Result} instance
}
