% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ret_err}
\alias{ret_err}
\title{Raise an Error but Return a Value}
\usage{
ret_err(x, msg)
}
\arguments{
\item{x}{value to return}

\item{msg}{error message to raise}
}
\value{
\code{x}, with an error
}
\description{
Raise an Error but Return a Value
}
\seealso{
Other helpers: 
\code{\link{add_n}()},
\code{\link{ret_null}()},
\code{\link{sleep_for}()},
\code{\link{square}()},
\code{\link{timestwo}()}
}
\concept{helpers}
