% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ret_null}
\alias{ret_null}
\title{Return \code{NULL}}
\usage{
ret_null()
}
\value{
\code{NULL}
}
\description{
Return \code{NULL}
}
\seealso{
Other helpers: 
\code{\link{add_n}()},
\code{\link{ret_err}()},
\code{\link{sleep_for}()},
\code{\link{square}()},
\code{\link{timestwo}()}
}
\concept{helpers}
