% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sleep_for}
\alias{sleep_for}
\title{Sleep but Return a Value}
\usage{
sleep_for(x, n)
}
\arguments{
\item{x}{value to return}

\item{n}{number of seconds to sleep for}
}
\value{
\code{x} after sleeping for \code{n} seconds
}
\description{
Sleep but Return a Value
}
\seealso{
Other helpers: 
\code{\link{add_n}()},
\code{\link{ret_err}()},
\code{\link{ret_null}()},
\code{\link{square}()},
\code{\link{timestwo}()}
}
\concept{helpers}
