% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{square}
\alias{square}
\title{Square a Number}
\usage{
square(x)
}
\arguments{
\item{x}{number to be squared}
}
\value{
\code{x^2}
}
\description{
Square a Number
}
\seealso{
Other helpers: 
\code{\link{add_n}()},
\code{\link{ret_err}()},
\code{\link{ret_null}()},
\code{\link{sleep_for}()},
\code{\link{timestwo}()}
}
\concept{helpers}
