% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{Timer}
\alias{Timer}
\title{Timer Monad}
\description{
Timer Monad

Timer Monad
}
\details{
Creates a \code{Timer} monad instance which can be used with the flatMap
operator \verb{\%>>=\%}. See the \verb{Monads in R} vignette for more details.

This container, when flatMapped, records the evaluation time of the
expressions used at each step, and these can be recovered from the object,
along with the expressions themselves.
}
\section{Usage}{

\preformatted{x <- Timer$new(value)

x$bind()

x$value
x$times
}
}

\section{Details}{

\code{$new(value)} creates a new instance wrapping \code{value}.

\code{$bind()} is used within \verb{\%>>=\%} and re-wraps as a new \code{Timer}.

\code{$value} extracts the \code{value} field. \code{value(x)} is a wrapper for this.
\code{$times} extracts the \code{times} field. \code{times(x)} is a wrapper for this.
}

\seealso{
\code{\link[=times]{times()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{value contained in the instance}

\item{\code{times}}{evaluation times}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimerMonad-new}{\code{Timer$new()}}
\item \href{#method-TimerMonad-bind}{\code{Timer$bind()}}
\item \href{#method-TimerMonad-clone}{\code{Timer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimerMonad-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimerMonad-new}{}}}
\subsection{Method \code{new()}}{
create a new \code{Timer} instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$new(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value to be contained}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{Timer} instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimerMonad-bind"></a>}}
\if{latex}{\out{\hypertarget{method-TimerMonad-bind}{}}}
\subsection{Method \code{bind()}}{
bind when using flatMap (\verb{\%>>=\%})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$bind(f, expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{function to apply}

\item{\code{expr}}{expression of the function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimerMonad-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimerMonad-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Timer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
