% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timerM}
\alias{timerM}
\title{Create a new \code{Timer} instance}
\usage{
timerM(value)
}
\arguments{
\item{value}{value to be contained}
}
\value{
a \code{Timer} monad instance
}
\description{
Create a new \code{Timer} instance
}
