% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{times}
\alias{times}
\title{Extract the times}
\usage{
times(x, digits = 6)
}
\arguments{
\item{x}{a \code{Timer} monad instance}

\item{digits}{minimum number of digits to print}
}
\value{
a \code{data.frame} of the commands and their timings
}
\description{
Extract the times
}
