% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestwo}
\alias{timestwo}
\title{Double a Number}
\usage{
timestwo(x)
}
\arguments{
\item{x}{number to be doubled}
}
\value{
\code{2*x}
}
\description{
Double a Number
}
\seealso{
Other helpers: 
\code{\link{add_n}()},
\code{\link{ret_err}()},
\code{\link{ret_null}()},
\code{\link{sleep_for}()},
\code{\link{square}()}
}
\concept{helpers}
