% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{Extract the value from a monad instance}
\usage{
value(x, ...)
}
\arguments{
\item{x}{an instance of a monad}

\item{...}{ignored}
}
\value{
the value wrapped by the monad
}
\description{
Extract the value from a monad instance
}
