% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{A list where all elements are the same class.}
\usage{
logger(value, log = character())
}
\arguments{
\item{value}{Value with logging context}

\item{log}{Character vector of logged values.}

\item{...}{Elements to combine into a list. All elements must be the
same type.}
}
\description{
A list where all elements are the same class.
}
\examples{
(l <- logger(5, "object created"))
(l2 <- l \%>>\% `*`(2))

set_value <- function(x, y) {
  logger(y, paste(x, "->", y))
}

l2 \%>+\% set_value(20)
}

