% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{maybe}
\alias{maybe}
\alias{nothing}
\title{Maybe there is something here?}
\usage{
maybe(x)

nothing()
}
\arguments{
\item{x}{Is there something here? Maybe.}
}
\description{
Or maybe there's just a \code{NULL}.
}
\examples{
double <- function(x) x * 2
maybe(10) \%>>\% double()
nothing() \%>>\% double()
}

